requirements.txt
aspose-cad==23.10
attrs==23.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
drawsvg==2.3.0
ezdxf==1.1.4
flasgger==0.9.7.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.47.0
greenlet==3.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mistune==3.0.2
numpy==1.26.2
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.1
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
six==1.16.0
SQLAlchemy==2.0.23
swig==4.1.1.post1
typing_extensions==4.9.0
Werkzeug==3.0.1
